% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQP_monitoring.R
\name{AQP_monitoring}
\alias{AQP_monitoring}
\title{Assess the Panel and Panelist Performance for QDA tests}
\usage{
AQP_monitoring(
  dataset,
  hidden_ref = "none",
  dif_ref = 10,
  incl_duplo = TRUE,
  duplicate,
  dif_duplo = 20,
  dif_range = 30,
  cor_lev = 0.7,
  panel_pmin = 0.05,
  panel_pmax = 0.05,
  panel_agrmin = 0.1,
  panel_agrmax = 0.2,
  sum_min = 50,
  sum_max = 75
)
}
\arguments{
\item{dataset}{Dataset to analyze}

\item{hidden_ref}{Name of the hidden reference sample (or 'none')}

\item{dif_ref}{Threshold for the difference between the reference and the hidden reference}

\item{incl_duplo}{Boolean: does the test includes duplicated samples?}

\item{duplicate}{Vector with the names of the two duplicates}

\item{dif_duplo}{Threshold for the difference between the duplicates}

\item{dif_range}{Threshold for the range between the samples}

\item{cor_lev}{Threshold for the correlation}

\item{panel_pmin}{Min pvalue for the panel discrimination}

\item{panel_pmax}{Max pvalue for the panel discrimination}

\item{panel_agrmin}{Min pvalue for the panel agreement}

\item{panel_agrmax}{Max pvalue for the panel agreement}

\item{sum_min}{Minimum value for the summary}

\item{sum_max}{Maximum value for the summary}
}
\value{
The different tables of results at the panel and panelist level, including summaries (per Judge and per Attribute)
}
\description{
Automated analysis for the panel and panelist performance in AQP test.
}
\examples{
NULL
}
