% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQP_plot.R
\name{AQP_plot}
\alias{AQP_plot}
\title{Visualize Results for AQP data}
\usage{
AQP_plot(dataset, hidden_ref = "none", duplicate = NULL)
}
\arguments{
\item{dataset}{Data to analyze}

\item{hidden_ref}{Name of the hidden reference}

\item{duplicate}{Names of the duplicate samples}
}
\value{
The Plot of interest
}
\description{
Visualize the scores provided by the panelists in AQP test.
}
\examples{
NULL
}
