% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR_monitoring.R
\name{CAR_monitoring}
\alias{CAR_monitoring}
\title{Assess the Panel and Panelist Performance for QDA tests}
\usage{
CAR_monitoring(
  dataset,
  reference = NULL,
  ref_dev = c(0.5, 1),
  incl_duplo = TRUE,
  duplicate = NULL,
  dup_dev = c(1, 2),
  dis_dev = c(2, 3),
  cor_val = c(0.7, 0.8),
  panel_pmin = 0.05,
  panel_pmax = 0.1,
  panel_agrmin = 0.1,
  panel_agrmax = 0.2,
  sum_min = 50,
  sum_max = 75
)
}
\arguments{
\item{dataset}{Dataset to analyze}

\item{reference}{Character: informs the hiddent reference sample}

\item{ref_dev}{Threshold for the deviation from the reference}

\item{incl_duplo}{Boolean: does the test includes duplicated samples?}

\item{duplicate}{Vector with the names of the two duplicates}

\item{dup_dev}{Threshold for the difference between the duplicates}

\item{dis_dev}{Threshold for the sample discrimination}

\item{cor_val}{Threshold for the agreement (correlation)}

\item{panel_pmin}{Min pvalue for the panel discrimination}

\item{panel_pmax}{Max pvalue for the panel discrimination}

\item{panel_agrmin}{Min pvalue for the panel agreement}

\item{panel_agrmax}{Max pvalue for the panel agreement}

\item{sum_min}{Minimum value for the summary}

\item{sum_max}{Maximum value for the summary}
}
\value{
The different tables of results at the panel and panelist level, including summaries (per Judge and per Attribute)
}
\description{
Assess the Panel and Panelist Performance for QDA tests
}
\examples{
NULL
}
