% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR_plot.R
\name{CAR_plot}
\alias{CAR_plot}
\title{Visualize Results for CAR data}
\usage{
CAR_plot(dataset, reference = NULL, duplicate = NULL)
}
\arguments{
\item{dataset}{Data to analyze}

\item{reference}{Name of the reference sample}

\item{duplicate}{Names of the duplicate samples}
}
\value{
The Plot of interest
}
\description{
Visualize the scores provided by the panelists in CAR test.
}
\examples{
NULL
}
