% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_anova.R
\name{cl_ANOVA}
\alias{cl_ANOVA}
\title{ANOVA per cluster}
\usage{
cl_ANOVA(dataset, test = "HSD")
}
\arguments{
\item{dataset}{The data to be analyzed}

\item{test}{The post'hoc test to be used ('LSD' or 'HSD')}
}
\value{
The ANOVA and post'hoc test results
}
\description{
Run the ANOVA with post'hoc analysis for each cluster separately
}
\examples{
NULL
}
