% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Perform HAC}
\usage{
clustering(
  dataset,
  center = TRUE,
  scale = TRUE,
  scale.unit = FALSE,
  consolidation = FALSE,
  nbcl,
  graph = TRUE,
  comparison = FALSE,
  stability = FALSE,
  NSim = 100
)
}
\arguments{
\item{dataset}{The data to analyse}

\item{center}{Boolean: should the consumer scores be centered (TRUE by default)}

\item{scale}{Boolean: should the consumer scores be standardized (TRUE by default)}

\item{scale.unit}{Boolean: should the PCA be standardized (FALSE by default)}

\item{consolidation}{Boolean: should the cluster solution obtained from HAC be consolidated using k-means (FALSE by default)}

\item{nbcl}{Integer: number of clusters to consider}

\item{graph}{Boolean: visualization of the dendrogram (TRUE by default)}

\item{comparison}{Boolean: compare the results before and after consolidation (FALSE by default)}

\item{stability}{Boolean: runs simulation to check for stability (FALSE by default)}

\item{NSim}{Integer: number of simulations to consider for the stability}
}
\value{
The different results of the clustering analysis
}
\description{
Performs Hierarchical Ascending Clustering by adding many interesting results
}
\examples{
NULL
}
