% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_code.R
\name{colour_code}
\alias{colour_code}
\title{Transforms Table into Corresponding Colour Coded Table}
\usage{
colour_code(data, value, direction)
}
\arguments{
\item{data}{Data to colour code}

\item{value}{Vector containing the set of 2 values to use for the colour coding to good/acceptable/bad}

\item{direction}{Integer: 1 means 'lower than min is good, and higher than max is bad'}
}
\value{
The corresponding table with the colour code.
}
\description{
Transforms Table into Corresponding Colour Coded Table
}
\examples{
NULL
}
