% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_filter.R
\name{consumer_filter}
\alias{consumer_filter}
\title{Filter Consumers}
\usage{
consumer_filter(dataset, range_filter = 2)
}
\arguments{
\item{dataset}{A table with the consumer data}

\item{range_filter}{Integer: the minimum range in 'Liking' scores consumer need to surpass to be kept in the dataset.}
}
\value{
The data table after removind undesired consumers
}
\description{
Define criteria to filter consumers with missing values or not enough variability in their 'Liking' scores
}
\examples{
NULL
}
