% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Converts values to colour code}
\usage{
convert(x, value1, value2)
}
\arguments{
\item{x}{Integer to convert}

\item{value1}{Integer for the lower limit}

\item{value2}{Integer for the higher limit}
}
\value{
Integer that corresponds to the colour code for x based on value1 and value2
}
\description{
Transform values into colour code based on pre-set limits: If x >= value1 -> 1, if x >= value2 --> 2, else 3
}
\examples{
NULL
}
