% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_correct.R
\name{design_correct}
\alias{design_correct}
\title{Correct for the Experimental Design}
\usage{
design_correct(dataset, pval = 0.05, correct = FALSE, seq_correct = NULL)
}
\arguments{
\item{dataset}{The data table to analyse}

\item{pval}{A p-value threshold to consider for significance (0.05 by default)}

\item{correct}{Boolean: FALSE by default, controls whether the liking score should be corrected.}

\item{seq_correct}{Vector: define which positions should be corrected (if correct = TRUE)}
}
\value{
The data after correction (if asked)
}
\description{
Correct the liking scores based on their position in the design
}
\examples{
NULL
}
