% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dol.R
\name{dol}
\alias{dol}
\title{Drivers of Liking}
\usage{
dol(data, attribute, pval_mod = 0.1, pval_quad = 0.1)
}
\arguments{
\item{data}{The data set to be analyzed}

\item{attribute}{Vector: contains the list of attributes to consider as drivers of liking}

\item{pval_mod}{Value: threshold for the p-value of the model}

\item{pval_quad}{Value: threshold to the p-value to keep quadratic effect}
}
\value{
The type of drivers of liking associated to each attribute
}
\description{
Run simple and quadratic regression for each attribute to evaluates the drivers of liking vs. Liking
}
\examples{
NULL
}
