% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCPC_CBI.R
\name{HCPC_CBI}
\alias{HCPC_CBI}
\title{Module to Extend Validation to HCPC}
\usage{
HCPC_CBI(dataset, k, consol = FALSE, scale.unit = FALSE)
}
\arguments{
\item{dataset}{The data to analyze}

\item{k}{Integer: the number of clusters}

\item{consol}{Boolean: should the cluster solution be consolidated using k-means}

\item{scale.unit}{Boolean: should the PCA be standardized or not}
}
\value{
The results of the bootstrap approach
}
\description{
Extend the bootstrap approach to check the validity to HCPC
}
\examples{
NULL
}
