% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDPref.R
\name{MDPref}
\alias{MDPref}
\title{Internal Preference Map (MDPref)}
\usage{
MDPref(dataset, scale.unit = TRUE, sensory = NULL)
}
\arguments{
\item{dataset}{The data table to be used to build the MDPred}

\item{scale.unit}{Boolean, TRUE by default that standardize the individual consumer liking scores}

\item{sensory}{The sensory data (should be the mean table with the product as first column and the attributes as subsequent columns)}
}
\value{
The results of the PrefMap (FactoMineR output)
}
\description{
Build the Internal Preference Map (or MDPref) by projecting as well the sensory data if avaialble.
}
\examples{
NULL
}
