% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODP_monitoring.R
\name{ODP_monitoring}
\alias{ODP_monitoring}
\title{Assess the Panel and Panelist Performance for ODP tests}
\usage{
ODP_monitoring(
  dataset,
  duplicate,
  lim_score = 20,
  lim_rank = 3,
  lim_discri = c(10, 20),
  lim_cor = 0.8,
  panel_pmin = 0.05,
  panel_pmax = 0.05,
  panel_agrmin = 0.1,
  panel_agrmax = 0.2,
  sum_min = 50,
  sum_max = 75
)
}
\arguments{
\item{dataset}{Dataset to analyze}

\item{duplicate}{Vector with the names of the two duplicates}

\item{lim_score}{Threshold for the difference between the scores of the duplicates}

\item{lim_rank}{Threshold for the difference between the ranks of the duplicates}

\item{lim_discri}{Threshold for the sample discrimination}

\item{lim_cor}{Threshold for the agreement (correlation)}

\item{panel_pmin}{Min pvalue for the panel discrimination}

\item{panel_pmax}{Max pvalue for the panel discrimination}

\item{panel_agrmin}{Min pvalue for the panel agreement}

\item{panel_agrmax}{Max pvalue for the panel agreement}

\item{sum_min}{Minimum value for the summary}

\item{sum_max}{Maximum value for the summary}
}
\value{
The different tables of results at the panel and panelist level, including summaries (per Judge and per Attribute)
}
\description{
Automated analysis for the panel and panelist performance in an ODP test. Both the intensity scores and the ranks are evaluated.
}
\examples{
NULL
}
