% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODP_plot.R
\name{ODP_plot}
\alias{ODP_plot}
\title{Visualize Results for ODP data}
\usage{
ODP_plot(dataset, duplicate)
}
\arguments{
\item{dataset}{Data to analyze}

\item{duplicate}{Names of the duplicate samples}
}
\value{
The Plot of interest
}
\description{
Visualize the scores provided by the panelists in ODP test.
}
\examples{
NULL
}
