% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist_agreement.R
\name{panellist_agreement}
\alias{panellist_agreement}
\title{Compute the Agreement between Panellists}
\usage{
panellist_agreement(
  dataset,
  firstvar = 6,
  lastvar = ncol(dataset),
  cor = "without",
  type = "Pearson"
)
}
\arguments{
\item{dataset}{Data frame to use}

\item{firstvar}{Integer: position of the first sensory attribute}

\item{lastvar}{Integer: position of the last sensory attribute}

\item{cor}{Character: 'full' or 'without' to consider or not the panellist in the panel scores}

\item{type}{Character: 'Pearson' or 'Spearman' to determine the type of correlation to consider}
}
\value{
The table of Correlation
}
\description{
Compute the agreement between the panellists through correlations (Pearson or Spearman). This correlation is measured against the entire panel or the panel without that panellist.
}
\examples{
NULL
}
