% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PrefMap.R
\name{plot.PrefMap}
\alias{plot.PrefMap}
\title{Surface Response Plot for PrefMap}
\usage{
\method{plot}{PrefMap}(
  x,
  space,
  map = "Proportion",
  colour = c("#39AECF", "white", "#DC3545"),
  col_sup = "grey32",
  include_var = FALSE,
  var_size = 1,
  var_col = "#F7B924",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  ...
)
}
\arguments{
\item{x}{Results obtained from the PrefMap function}

\item{space}{Sensory space}

\item{map}{Depends on the acceptance defined in PrefMap(): is it 'acceptance' or 'prediction'}

\item{colour}{The colouring scheme for the PrefMap (3 colours)}

\item{col_sup}{The colour for the supplementary individuals}

\item{include_var}{Boolean: should the variables be projected as well}

\item{var_size}{Integer: size the of the variable (if represented)}

\item{var_col}{Colour code to apply for the variables (if represented)}

\item{title}{Title of the graph}

\item{x_lab}{Title for the X-axis}

\item{y_lab}{Title for the Y-axis}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
The surface response plot
}
\description{
Build the surface response plot for the PrefMap
}
\examples{
NULL
}
