% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefMap.R
\name{PrefMap}
\alias{PrefMap}
\title{PrefMap}
\usage{
PrefMap(
  dataset,
  col.p = 2,
  col.j = 1,
  col.lik = 6,
  sensory = NULL,
  xlimits = NULL,
  ylimits = NULL,
  resolution = 200,
  regmod = "Best",
  param = "AdjR2",
  acceptance = "Mean",
  level = 0,
  predict = NULL,
  incProgress = NULL
)
}
\arguments{
\item{dataset}{The data to be analyzed}

\item{col.p}{Integer: corresponds to the position of the 'Product' column}

\item{col.j}{Integer: corresponds to the position of the 'Consumer' column}

\item{col.lik}{Integer: corresponds to the position of the 'Liking' column}

\item{sensory}{The sensory data - a table with the coordinates of the products}

\item{xlimits}{Vector: contains the min and max on the X-axis}

\item{ylimits}{Vector: contains the min and max on the Y-axis}

\item{resolution}{Integer: define the granularity for the prediction}

\item{regmod}{Type of model to chose from - 'Linear', 'Circular', 'Elliptic', 'Quadratic' and 'Best'}

\item{param}{The parameter to chose for selecting the best model - chose from 'AIC', 'BIC' and Adj. R2}

\item{acceptance}{The type of value to consider to define acceptance: 'mean' or 'value'}

\item{level}{The threshold to consider}

\item{predict}{Additional points to predict (e.g. supplementary entities)}

\item{incProgress}{Parameter to use for the progress bar...}
}
\value{
The numerical results of the PrefMap analysis
}
\description{
Run the PrefMap analysis
}
\examples{
NULL
}
