% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefMap_reg.R
\name{PrefMap_reg}
\alias{PrefMap_reg}
\title{Regression for PrefMap}
\usage{
PrefMap_reg(data, model, param, threshold, grid, indsup = NULL)
}
\arguments{
\item{data}{The data to be analyzed}

\item{model}{The model to consider - chose from 'Linear', 'Circular', 'Elliptic', 'Quadratic' and 'Best'}

\item{param}{The parameter to chose for selecting the best model - chose from 'AIC', 'BIC' and Adj. R2}

\item{threshold}{The p-value to consider for model selection}

\item{grid}{The values used for prediction for the PrefMap}

\item{indsup}{Information regarding supplementary individuals for extra calculation}
}
\value{
Prediction values and other results
}
\description{
The modeling tool for the Preference Mapping
}
\examples{
NULL
}
