% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{Cluster Profile Visualization}
\usage{
profiles(dataset)
}
\arguments{
\item{dataset}{The data to be represented}
}
\value{
The graphic visualizing the cluster profiles (in terms of preference) across samples
}
\description{
Generates the cluster profiles as a line-charts
}
\examples{
NULL
}
