% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA_monitoring.R
\name{QDA_monitoring}
\alias{QDA_monitoring}
\title{Assess the Panel and Panelist Performance for QDA tests}
\usage{
QDA_monitoring(
  dataset,
  col.r = "Replica",
  type = "reproducibility",
  random = TRUE,
  panel_sign = c(0.05, 0.1),
  panel_sd = c(10, 12),
  panellist_sign = c(0.05, 0.1),
  panellist_sd = c(10, 12),
  panellist_cor = c(0.5, 0.7),
  sum_min = 50,
  sum_max = 75,
  test = "QDA",
  hidden_ref = "none",
  dif_ref = c(5, 10),
  panel_agrmin = 0.1,
  panel_agrmax = 0.2,
  panel_cormin = 0.5,
  panel_cormax = 0.7
)
}
\arguments{
\item{dataset}{Dataset to analyze}

\item{col.r}{Integer: position of the replicate column}

\item{type}{Character: 'reproducibility' or 'repeatability' to define which model to use}

\item{random}{Boolean: should the panelist effect be considered as a random effect?}

\item{panel_sign}{Level of significance for the panel}

\item{panel_sd}{Threshold for the panel sd}

\item{panellist_sign}{Level of significance for the panelists}

\item{panellist_sd}{Threshold for the panelist sd (reproducibility)}

\item{panellist_cor}{Threshold for the panelist correlation (agreement)}

\item{sum_min}{Minimum value for the summary}

\item{sum_max}{Maximum value for the summary}

\item{test}{Type of test: 'QDA', 'AQP', or 'CAR'}

\item{hidden_ref}{Name of the hidden reference sample (or 'none')}

\item{dif_ref}{Threshold for the difference between the reference and the hidden reference}

\item{panel_agrmin}{Minimum value for the panel agreement}

\item{panel_agrmax}{Maximum value for the panel agreement}

\item{panel_cormin}{Minimum value for the panel correlation}

\item{panel_cormax}{Maximum value for the panel correlation}
}
\value{
The different tables of results at the panel and panelist level, including summaries (per Judge and per Attribute)
}
\description{
Automated analysis for the panel and panelist performance in quantitative descriptive analysis (QDA) when test is fully duplicated.
}
\examples{
NULL
}
