% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDA_plot.R
\name{QDA_plot}
\alias{QDA_plot}
\title{Build the Tucker Plot for QDA data}
\usage{
QDA_plot(
  dataset,
  firstvar = 6,
  lastvar = ncol(dataset),
  attr_show = "none",
  type_vis = "var"
)
}
\arguments{
\item{dataset}{Data to analyze}

\item{firstvar}{Integer: position of the first attribute}

\item{lastvar}{Integer: position of the last attribute}

\item{attr_show}{Vector containing the attribute to show}

\item{type_vis}{Character: 'ind' or 'var'}
}
\value{
The Plot of interest
}
\description{
Visualize the agreement between panelist for the different attributes in a PCA plot.
}
\examples{
NULL
}
