% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversereg.R
\name{reversereg}
\alias{reversereg}
\title{Reverse Regression}
\usage{
reversereg(coord, senso, coord.ideal)
}
\arguments{
\item{coord}{The coordinates of the products on the PrefMap space}

\item{senso}{The sensory profiles of the samples}

\item{coord.ideal}{The coordinates of the sample profile to predict}
}
\value{
The predicted sensory profile of the optimal sample
}
\description{
Reverse regression is used to predict the optimal sample from the PrefMap space by providing the coordinates of the optimal sample to predict.
}
\examples{
NULL
}
