% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot.R
\name{spiderplot}
\alias{spiderplot}
\title{Spider Plot}
\usage{
spiderplot(data, var)
}
\arguments{
\item{data}{The data to be analyzed}

\item{var}{The list of attributes in the data}
}
\value{
The spider plot
}
\description{
Build a spider plot to compare sensory profiles of samples
}
\examples{
NULL
}
