% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tab.R
\name{split_tab}
\alias{split_tab}
\title{Split Tables}
\usage{
split_tab(data)
}
\arguments{
\item{data}{Full data to be split}
}
\value{
The 2 sets of table
}
\description{
Split tables into its value component and color (column ends with '_col') component.
}
\examples{
NULL
}
