% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tab.R
\name{summary_tab}
\alias{summary_tab}
\title{Build the Summary Tables}
\usage{
summary_tab(data, type = "Judge")
}
\arguments{
\item{data}{Data to use}

\item{type}{Character: 'Judge' or 'Attribute'}
}
\value{
A summary table for either the Judge or the Attribute
}
\description{
Build the summary tables for the different abilities.
}
\examples{
NULL
}
