% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trimming}
\alias{trimming}
\title{Cluster Trimming}
\usage{
trimming(dataset, min_cor = 0.2)
}
\arguments{
\item{dataset}{The data to be trimmed}

\item{min_cor}{Value: the minimum correlation to consider to keep the consumer in the analysis (use -1 to keep all consumers)}
}
\value{
The data after trimming
}
\description{
The cluster results can be trimmed by removing consumers who are not correlating well with their cluster means.
}
\examples{
NULL
}
